restaurado de funcionalidades
