+++ src/app/fabric/page.tsx
@@ -1,6 +1,17 @@
-import Fabric from "@/components/fabric/fabric";
+"use client";
+import { useEffect, useState } from "react";
+import { Fabric } from "@/components/fabric/fabric";
 
 export default function FabricPage() {
+  const [isMobile, setIsMobile] = useState<null | boolean>(null);
 
-  return <Fabric />;
+  useEffect(() => {
+    if (typeof window === "undefined") return;
+    setIsMobile(/Mobi|Android/i.test(navigator.userAgent));
+  }, []);
+
+  // Evitamos renderizar hasta que el estado se resuelva
+  if (isMobile === null) return null;
+
+  return <Fabric isMobile={isMobile} />;
 }
