--- src/components/fabric/deviceDetection.ts
@@ -1,4 +1,0 @@
-// deviceDetection.ts
-export const isMobile = /Mobi|Android/i.test(
-  typeof navigator !== "undefined" ? navigator.userAgent : ""
-);
