--- src/components/fabric/fabricHelpers.ts
@@ -1,11 +1,0 @@
-// fabricHelpers.ts
-export const sendTextToFront = (fabricCanvas: any) => {
-  if (!fabricCanvas) return;
-
-  fabricCanvas.getObjects().forEach((obj: any) => {
-    if (obj.type === "textbox") {
-      fabricCanvas.bringToFront(obj);
-    }
-  });
-  fabricCanvas.renderAll();
-};
